moving TokenMigration screen to it's own folder
